xmhf: shared/shared-archds.c: change GDT and IDT descriptors to have a 64-bit base address; types appropriately revised for 64-bits.

+++ include/arch/x86-vmx-x86pc/xmhf-x86-vmx-x86pc.h
@@ -145,16 +145,30 @@ struct _memorytype {
 
 
 //x86 GDT descriptor type
+//typedef struct {
+//		u16 size;
+//		u32 base;
+//} __attribute__((packed)) arch_x86_gdtdesc_t;
+
+//x86 IDT descriptor type
+//typedef struct {
+//		u16 size;
+//		u32 base;
+//} __attribute__((packed)) arch_x86_idtdesc_t;
+
+
+//x86 GDT descriptor type
 typedef struct {
 		u16 size;
-		u32 base;
+		u64 base;
 } __attribute__((packed)) arch_x86_gdtdesc_t;
 
 //x86 IDT descriptor type
 typedef struct {
 		u16 size;
-		u32 base;
+		u64 base;
 } __attribute__((packed)) arch_x86_idtdesc_t;
+
 
 //TSS descriptor (partial)
 typedef struct __tss {
